requirements.txt
python-dotenv>=1.0.0
discord-ext-ipcx>=0.1.1
quart>=0.19.3
hypercorn[uvloop]>=0.14.4
